Add unit route POST
